Split sc.h into multiple headers.
