+++ compiler/errors.cpp
@@ -37,7 +37,10 @@
   #include <alloc/fortify.h>
 #endif
 #include "sc.h"
+#include "scvars.h"
 #include "errors.h"
+#include "sclist.h"
+#include "libpawnc.h"
 
 #if defined _MSC_VER
   #pragma warning(push)
