+++ compiler/sc.h
@@ -65,11 +65,6 @@
 #define sDEF_PREFIX   "sourcemod.inc" /* default prefix filename */
 #define sTAGS_MAX		16  /* maximum number of tags on an argument */
 
-typedef union {
-  void *pv;                 /* e.g. a name */
-  int i;
-} stkitem;                  /* type of items stored on the compiler stack */
-
 struct arginfo {  /* function argument info */
   char name[sNAMEMAX+1];
   char ident;           /* iVARIABLE, iREFERENCE, iREFARRAY or iVARARGS */
@@ -378,14 +373,6 @@ enum {
   statSKIP,     /* skipping output */
 };
 
-struct stringlist {
-  stringlist *next;
-  union {
-    char *line;
-    stringlist *tail;
-  };
-};
-
 typedef struct s_stringpair {
   struct s_stringpair *next;
   char *first;
@@ -394,183 +381,6 @@ typedef struct s_stringpair {
   char flags;
   char *documentation;
 } stringpair;
-
-// Helper for token info.
-typedef struct {
-  int id;
-  cell val;
-  char *str;
-} token_t;
-
-// The method name buffer is larger since we can include our parent class's
-// name, a "." to separate it, a "~" for constructors, or a ".get/.set" for
-// accessors.
-#define METHOD_NAMEMAX sNAMEMAX * 2 + 6
-
-typedef struct {
-  token_t tok;
-  char name[METHOD_NAMEMAX + 1];
-} token_ident_t;
-
-/* macros for code generation */
-#define opcodes(n)      ((n)*sizeof(cell))      /* opcode size */
-#define opargs(n)       ((n)*sizeof(cell))      /* size of typical argument */
-
-/*  Tokens recognized by lex()
- *  Some of these constants are assigned as well to the variable "lastst" (see SC1.C)
- */
-enum TokenKind {
-  /* value of first multi-character operator */
-  tFIRST     = 256,
-  /* multi-character operators */
-  taMULT     = tFIRST, /* *= */
-  taDIV,     /* /= */
-  taMOD,     /* %= */
-  taADD,     /* += */
-  taSUB,     /* -= */
-  taSHL,     /* <<= */
-  taSHRU,    /* >>>= */
-  taSHR,     /* >>= */
-  taAND,     /* &= */
-  taXOR,     /* ^= */
-  taOR,      /* |= */
-  tlOR,      /* || */
-  tlAND,     /* && */
-  tlEQ,      /* == */
-  tlNE,      /* != */
-  tlLE,      /* <= */
-  tlGE,      /* >= */
-  tSHL,      /* << */
-  tSHRU,     /* >>> */
-  tSHR,      /* >> */
-  tINC,      /* ++ */
-  tDEC,      /* -- */
-  tELLIPS,   /* ... */
-  tDBLDOT,   /* .. */
-  tDBLCOLON, /* :: */
-  /* value of last multi-character operator */
-  tMIDDLE    = tDBLCOLON,
-/* reserved words (statements) */
-  tACQUIRE,
-  tAS,
-  tASSERT,
-  tBREAK,
-  tBUILTIN,
-  tCATCH,
-  tCASE,
-  tCAST_TO,
-  tCELLSOF,
-  tCHAR,
-  tCONST,
-  tCONTINUE,
-  tDECL,
-  tDEFAULT,
-  tDEFINED,
-  tDELETE,
-  tDO,
-  tDOUBLE,
-  tELSE,
-  tENUM,
-  tEXIT,
-  tEXPLICIT,
-  tFINALLY,
-  tFOR,
-  tFOREACH,
-  tFORWARD,
-  tFUNCENUM,
-  tFUNCTAG,
-  tFUNCTION,
-  tGOTO,
-  tIF,
-  tIMPLICIT,
-  tIMPORT,
-  tIN,
-  tINT,
-  tINT8,
-  tINT16,
-  tINT32,
-  tINT64,
-  tINTERFACE,
-  tINTN,
-  tLET,
-  tMETHODMAP,
-  tNAMESPACE,
-  tNATIVE,
-  tNEW,
-  tNULL,
-  tNULLABLE,
-  tOBJECT,
-  tOPERATOR,
-  tPACKAGE,
-  tPRIVATE,
-  tPROTECTED,
-  tPUBLIC,
-  tREADONLY,
-  tRETURN,
-  tSEALED,
-  tSIZEOF,
-  tSLEEP,
-  tSTATIC,
-  tSTOCK,
-  tSTRUCT,
-  tSWITCH,
-  tTAGOF,
-  tTHIS,
-  tTHROW,
-  tTRY,
-  tTYPEDEF,
-  tTYPEOF,
-  tTYPESET,
-  tUINT8,
-  tUINT16,
-  tUINT32,
-  tUINT64,
-  tUINTN,
-  tUNION,
-  tUSING,
-  tVAR,
-  tVARIANT,
-  tVIEW_AS,
-  tVIRTUAL,
-  tVOID,
-  tVOLATILE,
-  tWHILE,
-  tWITH,
-  /* compiler directives */
-  tpASSERT,     /* #assert */
-  tpDEFINE,
-  tpELSE,       /* #else */
-  tpELSEIF,     /* #elseif */
-  tpENDIF,
-  tpENDINPUT,
-  tpENDSCRPT,
-  tpERROR,
-  tpWARNING,
-  tpFILE,
-  tpIF,         /* #if */
-  tINCLUDE,
-  tpLINE,
-  tpPRAGMA,
-  tpTRYINCLUDE,
-  tpUNDEF,
-  tLAST = tpUNDEF,   /* value of last multi-character match-able token */
-  /* semicolon is a special case, because it can be optional */
-  tTERM,          /* semicolon or newline */
-  tENDEXPR,       /* forced end of expression */
-  /* other recognized tokens */
-  tNUMBER,        /* integer number */
-  tRATIONAL,      /* rational number */
-  tSYMBOL,
-  tLABEL,
-  tSTRING,
-  tPENDING_STRING, /* string, but not yet dequeued */
-  tEXPR,          /* for assigment to "lastst" only (see SC1.C) */
-  tENDLESS,       /* endless loop, for assigment to "lastst" only */
-  tEMPTYBLOCK,    /* empty blocks for AM bug 4825 */
-  tEOL,           /* newline, only returned by peek_new_line() */
-  tNEWDECL,       /* for declloc() */
-  tLAST_TOKEN_ID
-};
 
 /* (reversed) evaluation of staging buffer */
 #define sSTARTREORDER 0x01
@@ -605,29 +415,9 @@ enum TokenKind {
 #define sEXPRMARK       2       /* mark start of expression */
 #define sEXPRRELEASE    3       /* mark end of expression */
 
-enum {
-  sOPTIMIZE_NONE,               /* no optimization */
-  sOPTIMIZE_NOMACRO,            /* no macro instructions */
-  sOPTIMIZE_DEFAULT,            /* full optimization */
-  /* ----- */
-  sOPTIMIZE_NUMBER
-};
-
-typedef enum s_regid {
-  sPRI,                         /* indicates the primary register */
-  sALT,                         /* indicates the secundary register */
-} regid;
-
-typedef enum s_optmark {
-  sEXPR,                        /* end of expression (for expressions that form a statement) */
-  sPARM,                        /* end of parameter (in a function parameter list) */
-  sLDECL,                       /* start of local declaration (variable) */
-} optmark;
-
-#define suSLEEP_INSTR 0x01      /* the "sleep" instruction was used */
-
 #define CELL_MAX      (((ucell)1 << (sizeof(cell)*8-1)) - 1)
 
+struct token_t;
 
 /*
  * Functions you call from the "driver" program
@@ -636,38 +426,10 @@ int pc_compile(int argc, char **argv);
 int pc_addconstant(const char *name,cell value,int tag);
 int pc_addtag(const char *name);
 int pc_findtag(const char *name);
-int pc_enablewarning(int number,int enable);
 const char *pc_tagname(int tag);
 int parse_decl(declinfo_t *decl, int flags);
 const char *type_to_name(int tag);
 bool parse_new_typename(const token_t *tok, int *tagp);
-
-/*
- * Functions called from the compiler (to be implemented by you)
- */
-
-/* general console output */
-int pc_printf(const char *message,...);
-
-/* input from source file */
-void *pc_opensrc(char *filename); /* reading only */
-void *pc_createsrc(char *filename);
-void pc_closesrc(void *handle);   /* never delete */
-char *pc_readsrc(void *handle,unsigned char *target,int maxchars);
-int pc_writesrc(void *handle,unsigned char *source);
-void *pc_getpossrc(void *handle,void *position); /* mark the current position */
-void pc_resetsrc(void *handle,void *position);  /* reset to a position marked earlier */
-int  pc_eofsrc(void *handle);
-
-/* output to intermediate (.ASM) file */
-void *pc_openasm(char *filename); /* read/write */
-void pc_closeasm(void *handle,int deletefile);
-void pc_resetasm(void *handle);
-int  pc_writeasm(void *handle,const char *str);
-char *pc_readasm(void *handle,char *target,int maxchars);
-
-void sp_fdbg_ntv_start(int num_natives);
-void sp_fdbg_ntv_hook(int index, symbol *sym);
 
 /* function prototypes in SC1.C */
 void set_extension(char *filename,const char *extension,int force);
@@ -681,312 +443,12 @@ void delete_consttable(constvalue *table);
 symbol *add_constant(const char *name,cell val,int vclass,int tag);
 int get_actual_compound(symbol *sym);
 
-/* function prototypes in SC2.C */
-#define PUSHSTK_P(v)  { stkitem s_; s_.pv=(v); pushstk(s_); }
-#define PUSHSTK_I(v)  { stkitem s_; s_.i=(v); pushstk(s_); }
-#define POPSTK_P()    (popstk().pv)
-#define POPSTK_I()    (popstk().i)
-void pushstk(stkitem val);
-stkitem popstk(void);
-void clearstk(void);
-int plungequalifiedfile(char *name);  /* explicit path included */
-int plungefile(char *name,int try_currentpath,int try_includepaths);   /* search through "include" paths */
-void preprocess(void);
-void lexinit(void);
-int lex(cell *lexvalue,char **lexsym);
-int lextok(token_t *tok);
-int lexpeek(int id);
-void lexpush(void);
-void lexclr(int clreol);
-int matchtoken(int token);
-int tokeninfo(cell *val,char **str);
-int needtoken(int token);
-int matchtoken2(int id, token_t *tok);
-int expecttoken(int id, token_t *tok);
-int matchsymbol(token_ident_t *ident);
-int needsymbol(token_ident_t *ident);
-int peek_same_line();
-int require_newline(int allow_semi);
-void litadd(cell value);
-void litinsert(cell value,int pos);
-int alphanum(char c);
-int ishex(char c);
-int isoctal(char c);
-void delete_symbol(symbol *root,symbol *sym);
-void delete_symbols(symbol *root,int level,int del_labels,int delete_functions);
-int refer_symbol(symbol *entry,symbol *bywhom);
-void markusage(symbol *sym,int usage);
-symbol *findglb(const char *name);
-symbol *findloc(const char *name);
-symbol *findconst(const char *name,int *matchtag);
-symbol *addsym(const char *name,cell addr,int ident,int vclass,int tag, int usage);
-symbol *addvariable(const char *name,cell addr,int ident,int vclass,int tag,
-                            int dim[],int numdim,int idxtag[]);
-symbol *addvariable2(const char *name,cell addr,int ident,int vclass,int tag,
-                             int dim[],int numdim,int idxtag[],int slength);
-symbol *addvariable3(declinfo_t *decl,cell addr,int vclass,int slength);
-int getlabel(void);
-char *itoh(ucell val);
-
-#define MATCHTAG_COERCE       0x1 // allow coercion
-#define MATCHTAG_SILENT       0x2 // silence the error(213) warning
-#define MATCHTAG_COMMUTATIVE  0x4 // order does not matter
-#define MATCHTAG_DEDUCE       0x8 // correct coercion
-
-/* function prototypes in SC3.C */
-int check_userop(void (*oper)(void),int tag1,int tag2,int numparam,
-                         value *lval,int *resulttag);
-int matchtag(int formaltag,int actualtag,int allowcoerce);
-int expression(cell *val,int *tag,symbol **symptr,int chkfuncresult,value *_lval);
-cell array_totalsize(symbol *sym);
-int matchtag_string(int ident, int tag);
-int checkval_string(value *sym1, value *sym2);
-int checktag_string(int tag, value *sym1);
-int lvalexpr(svalue *sval);
-
-/* function prototypes in SC4.C */
-void writeleader(symbol *root);
-void writetrailer(void);
-void begcseg(void);
-void begdseg(void);
-void setline(int chkbounds);
-void setfiledirect(char *name);
-void setlinedirect(int line);
-void setlabel(int index);
-void markexpr(optmark type,const char *name,cell offset);
-void startfunc(const char *fname);
-void endfunc(void);
-void rvalue(value *lval);
-void rvalue(svalue *sval);
-void address(symbol *ptr,regid reg);
-void store(value *lval);
-void loadreg(cell address,regid reg);
-void storereg(cell address,regid reg);
-void memcopy(cell size);
-void copyarray(symbol *sym,cell size);
-void fillarray(symbol *sym,cell size,cell value);
-void ldconst(cell val,regid reg);
-void moveto1(void);
-void move_alt(void);
-void pushreg(regid reg);
-void pushval(cell val);
-void popreg(regid reg);
-void genarray(int dims, int _autozero);
-void swap1(void);
-void ffswitch(int label);
-void ffcase(cell value,char *labelname,int newtable);
-void ffcall(symbol *sym,const char *label,int numargs);
-void ffret();
-void ffabort(int reason);
-void ffbounds(cell size);
-void ffbounds();
-void jumplabel(int number);
-void defstorage(void);
-void modstk(int delta);
-void modheap(int delta);
-void modheap_i();
-void setheap_pri(void);
-void setheap(cell value);
-void cell2addr(void);
-void cell2addr_alt(void);
-void char2addr(void);
-void addconst(cell value);
-void setheap_save(cell value);
-void stradjust(regid reg);
-void invoke_getter(methodmap_method_t *method);
-void invoke_setter(methodmap_method_t *method, int save);
-void inc_pri();
-void dec_pri();
-void load_hidden_arg();
-void load_glbfn(symbol *sym);
-
-/*  Code generation functions for arithmetic operators.
- *
- *  Syntax: o[u|s|b]_name
- *          |   |   | +--- name of operator
- *          |   |   +----- underscore
- *          |   +--------- "u"nsigned operator, "s"igned operator or "b"oth
- *          +------------- "o"perator
- */
-void os_mult(void); /* multiplication (signed) */
-void os_div(void);  /* division (signed) */
-void os_mod(void);  /* modulus (signed) */
-void ob_add(void);  /* addition */
-void ob_sub(void);  /* subtraction */
-void ob_sal(void);  /* shift left (arithmetic) */
-void os_sar(void);  /* shift right (arithmetic, signed) */
-void ou_sar(void);  /* shift right (logical, unsigned) */
-void ob_or(void);   /* bitwise or */
-void ob_xor(void);  /* bitwise xor */
-void ob_and(void);  /* bitwise and */
-void ob_eq(void);   /* equality */
-void ob_ne(void);   /* inequality */
-void relop_prefix(void);
-void relop_suffix(void);
-void os_le(void);   /* less or equal (signed) */
-void os_ge(void);   /* greater or equal (signed) */
-void os_lt(void);   /* less (signed) */
-void os_gt(void);   /* greater (signed) */
-
-void lneg(void);
-void neg(void);
-void invert(void);
-void nooperation(void);
-void inc(value *lval);
-void dec(value *lval);
-void jmp_ne0(int number);
-void jmp_eq0(int number);
-void outval(cell val,int newline);
-
-/* function prototypes in SC5.C */
-int error(int number,...);
-int error(symbol* sym, int number, ...);
-void errorset(int code,int line);
-
-/* function prototypes in SC6.C */
-void assemble(const char *outname, void *fin);
-
-/* function prototypes in SC7.C */
-void stgbuffer_cleanup(void);
-void stgmark(char mark);
-void stgwrite(const char *st);
-void stgout(int index);
-void stgdel(int index,cell code_index);
-int stgget(int *index,cell *code_index);
-void stgset(int onoff);
-int phopt_init(void);
-int phopt_cleanup(void);
-
-/* function prototypes in SCLIST.C */
-char* duplicatestring(const char* sourcestring);
-stringpair *insert_alias(const char *name,const char *alias);
-stringpair *find_alias(char *name);
-int lookup_alias(char *target,const char *name);
-void delete_aliastable(void);
-stringlist *insert_path(char *path);
-char *get_path(int index);
-void delete_pathtable(void);
-stringpair *insert_subst(const char *pattern,const char *substitution,int prefixlen);
-int get_subst(int index,char **pattern,char **substitution);
-stringpair *find_subst(char *name,int length);
-int delete_subst(char *name,int length);
-void delete_substtable(void);
-stringlist *insert_sourcefile(char *string);
-char *get_sourcefile(int index);
-void delete_sourcefiletable(void);
-stringlist *insert_inputfile(char *string);
-char *get_inputfile(int index);
-void delete_inputfiletable(void);
-stringlist *insert_docstring(char *string);
-char *get_docstring(int index);
-void delete_docstring(int index);
-void delete_docstringtable(void);
-stringlist *insert_autolist(const char *string);
-char *get_autolist(int index);
-void delete_autolisttable(void);
-stringlist *insert_dbgfile(const char *filename);
-stringlist *insert_dbgline(int linenr);
-stringlist *insert_dbgsymbol(symbol *sym);
-char *get_dbgstring(int index);
-void delete_dbgstringtable(void);
-stringlist *get_dbgstrings();
-void delete_stringtable(stringlist *root);
-
-/* function prototypes in SCI18N.C */
-cell get_utf8_char(const unsigned char *string,const unsigned char **endptr);
-int scan_utf8(void *fp,const char *filename);
-
-/* external variables (defined in scvars.c) */
-#if !defined SC_SKIP_VDECL
-typedef struct HashTable HashTable;
-extern struct HashTable *sp_Globals;
-extern symbol loctab;       /* local symbol table */
-extern symbol glbtab;       /* global symbol table */
-extern cell *litq;          /* the literal queue */
-extern unsigned char pline[]; /* the line read from the input file */
-extern const unsigned char *lptr;/* points to the current position in "pline" */
-extern constvalue libname_tab;/* library table (#pragma library "..." syntax) */
-extern int pc_addlibtable;  /* is the library table added to the AMX file? */
-extern symbol *curfunc;     /* pointer to current function */
-extern char *inpfname;      /* name of the file currently read from */
-extern char outfname[];     /* intermediate (assembler) file name */
-extern char binfname[];     /* binary file name */
-extern char errfname[];     /* error file name */
-extern char sc_ctrlchar;    /* the control character (or escape character) */
-extern char sc_ctrlchar_org;/* the default control character */
-extern int litidx;          /* index to literal table */
-extern int litmax;          /* current size of the literal table */
-extern int stgidx;          /* index to the staging buffer */
-extern int sc_labnum;       /* number of (internal) labels */
-extern int staging;         /* true if staging output */
-extern cell declared;       /* number of local cells declared */
-extern cell glb_declared;   /* number of global cells declared */
-extern cell code_idx;       /* number of bytes with generated code */
-extern int ntv_funcid;      /* incremental number of native function */
-extern int errnum;          /* number of errors */
-extern int warnnum;         /* number of warnings */
-extern int sc_debug;        /* debug/optimization options (bit field) */
-extern int sc_packstr;      /* strings are packed by default? */
-extern int sc_asmfile;      /* create .ASM file? */
-extern int sc_listing;      /* create .LST file? */
-extern int sc_needsemicolon;/* semicolon required to terminate expressions? */
-extern int sc_dataalign;    /* data alignment value */
-extern int sc_showincludes; /* show include files? */
-extern int curseg;          /* 1 if currently parsing CODE, 2 if parsing DATA */
-extern cell pc_stksize;     /* stack size */
-extern int freading;        /* is there an input file ready for reading? */
-extern int fline;           /* the line number in the current file */
-extern short fnumber;       /* number of files in the input file table */
-extern short fcurrent;      /* current file being processed */
-extern short sc_intest;     /* true if inside a test */
-extern int sideeffect;      /* true if an expression causes a side-effect */
-extern int stmtindent;      /* current indent of the statement */
-extern int indent_nowarn;   /* skip warning "217 loose indentation" */
-extern int sc_tabsize;      /* number of spaces that a TAB represents */
-extern short sc_allowtags;  /* allow/detect tagnames in lex() */
-extern int sc_status;       /* read/write status */
-extern int sc_err_status;   /* TRUE if errors should be generated even if sc_status = SKIP */
-extern int sc_rationaltag;  /* tag for rational numbers */
-extern int rational_digits; /* number of fractional digits */
-extern short sc_is_utf8;    /* is this source file in UTF-8 encoding */
-extern char *pc_deprecate;  /* if non-NULL, mark next declaration as deprecated */
-extern int pc_optimize;     /* (peephole) optimization level */
-extern int pc_memflags;     /* special flags for the stack/heap usage */
-extern int pc_functag;      /* global function tag */
-extern int pc_tag_string;   /* global String tag */
-extern int pc_tag_void;     /* global void tag */
-extern int pc_tag_object;   /* root object tag */
-extern int pc_tag_bool;     /* global bool tag */
-extern int pc_tag_null_t;   /* the null type */
-extern int pc_tag_nullfunc_t;   /* the null function type */
-extern int pc_anytag;       /* global any tag */
-extern int glbstringread;	  /* last global string read */
-extern int sc_require_newdecls; /* only newdecls are allowed */
-extern bool sc_warnings_are_errors;
-extern unsigned sc_total_errors;
-extern int pc_code_version; /* override the code version */
-
-// Returns true if compilation is in its second phase (writing phase) and has
-// so far proceeded without error.
-static inline bool cc_ok()
-{
-  return sc_status == statWRITE && sc_total_errors == 0;
-}
-
-extern void *inpf;          /* file read from (source or include) */
-extern void *inpf_org;      /* main source file */
-extern void *outf;          /* file written to */
-
-extern jmp_buf errbuf;      /* target of longjmp() on a fatal error */
-
 #if defined WIN32
 # if !defined snprintf
 #  define snprintf _snprintf
 #  define vsnprintf _vsnprintf
 # endif
 #endif
-
-#endif /* SC_SKIP_VDECL */
 
 typedef struct array_info_s
 {
@@ -998,27 +460,6 @@ typedef struct array_info_s
   int *cur_dims;					/* Current dimensions the recursion is at */
   cell *base;						/* &litq[startlit] */
 } array_info_t;
-
-enum FatalError {
-  FIRST_FATAL_ERROR = 183,
-
-  FATAL_ERROR_READ  = FIRST_FATAL_ERROR,
-  FATAL_ERROR_WRITE,
-  FATAL_ERROR_ALLOC_OVERFLOW,
-  FATAL_ERROR_OOM,
-  FATAL_ERROR_INVALID_INSN,
-  FATAL_ERROR_INT_OVERFLOW,
-  FATAL_ERROR_SCRIPT_OVERFLOW,
-  FATAL_ERROR_OVERWHELMED_BY_BAD,
-  FATAL_ERROR_NO_CODEPAGE,
-  FATAL_ERROR_INVALID_PATH,
-  FATAL_ERROR_ASSERTION_FAILED,
-  FATAL_ERROR_USER_ERROR,
-  FATAL_ERROR_NO_GENERATED_CODE,
-  FATAL_ERROR_FUNCENUM,
-
-  FATAL_ERRORS_TOTAL
-};
 
 struct AutoDisableLiteralQueue
 {
