+++ compiler/sctracker.h
@@ -2,6 +2,9 @@
 #ifndef _INCLUDE_SOURCEPAWN_COMPILER_TRACKER_H_
 #define _INCLUDE_SOURCEPAWN_COMPILER_TRACKER_H_
 
+#include "scvars.h"
+#include "lexer.h"
+
 #define MEMUSE_STATIC      0
 #define MEMUSE_DYNAMIC     1
